requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cPython==0.0.6
cycler==0.10.0
Cython==0.29.22
Django==3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
future==0.18.2
fvcore==0.1.3.post20210305
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
iopath==0.1.4
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.20.1
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.2
Pillow==8.1.0
portalocker==2.2.1
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
PyJWT==1.7.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
utils==1.0.1
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8